update all file

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -31,7 +31,7 @@ public class JournalEntryService {
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntry().add(saved);
             user.setUserName(null);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             System.out.print(e);
             throw new RuntimeException("An error occurred while saving the entry.", e);
@@ -50,11 +50,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntry().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean remove = false;
+        try {
+            User user = userService.findByUserName(userName);
+            remove = user.getJournalEntry().removeIf(x -> x.getId().equals(id));
+            if (remove) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch(Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return remove;
     }
 }
